+++ project.flutter/client_app/lib/data/library.dart
@@ -18,6 +18,7 @@ enum MyLibrarySort {
 enum MyLibraryGrid {
 	two,
 	three,
+	list,
 }
 
 /// A class to manage the library of programs.